Initializes project

`org::initialize_project` takes in 2+ arguments. It then saves its results (i.e. folder locations) in `org::project`, which you will use in all of your subsequent code.

initialize_project(
  home = NULL,
  results = NULL,
  folders_to_be_sourced = "R",
  source_folders_absolute = FALSE,
  create_folders = FALSE,
  silent = FALSE,
  encode_from = "UTF-8",
  encode_to = "latin1",
  ...
)

Arguments

home
The folder containing 'Run.R' and 'R/'

results
A folder inside `results` with today's date will be created and it will be accessible via `org::project$results_today` (this is where you will store all of your results)

folders_to_be_sourced
The names of folders that live inside `home` and all .r and .R files inside it will be sourced into the global environment.

source_folders_absolute
If `TRUE` then `folders_to_be_sourced` is an absolute folder reference. If `FALSE` then `folders_to_be_sourced` is relative and inside `home`.

create_folders
Recommended that this is set to `TRUE`. It allows `org` to create any folders that are missing.

silent
Silence all feedback

encode_from
Folders current encoding (only used on Windows)

encode_to
Folders final encoding (only used on Windows)

...
Other folders that you would like to reference

Details

You need to set `create_folders=TRUE` for this function to create today's folder (org::project$results_today).

For more details see the help vignette: vignette("intro", package = "org")

Examples

if (FALSE) {
org::initialize(
  home = "/git/analyses/2019/analysis3/",
  results = "/dropbox/analyses_results/2019/analysis3/",
  raw = "/data/analyses/2019/analysis3/"
)
org::project$results_today
org::project$raw
}